Refactor to take into account that we have more platforms supported now
